Moved data containers from ffmpeg_reader.py into separate files
