+++ moviepy/video/io/ffmpeg_reader_utils/processing_state.py
@@ -1,0 +1,5 @@
+class ProcessingState:
+    def __init__(self, proc=None, pos=0, lastread=None):
+        self.proc = proc
+        self.pos = pos
+        self.lastread = lastread
